FITS changes to support better floating point images and using floating point darks and flats with 16 bit integer data images.

+++ Tangra 3/Helpers/FITSHelper.cs
@@ -210,7 +210,7 @@ namespace Tangra.Helpers
 					dataRow = new short[arr.Length];
 					for (int i = 0; i < arr.Length; i++)
 					{
-						dataRow[i] = Convert.ToInt16(arr.GetValue(i));
+						dataRow[i] = (short)Convert.ToInt32(arr.GetValue(i));
 					}
 				}
 				else
